01enWSL/pyproject.toml
# pyproject.toml — Week 1 Laboratory Kit Configuration
# NETWORKING class — ASE, CSIE | by ing. dr. Antonio Clim

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "week1-network-fundamentals"
version = "1.1.0"
description = "Week 1 Laboratory Kit: Network Fundamentals - Computer Networks Course"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "ing. dr. Antonio Clim"}
]
maintainers = [
    {name = "ing. dr. Antonio Clim"}
]
keywords = [
    "networking",
    "education",
    "tcp-ip",
    "sockets",
    "wireshark",
    "docker",
    "laboratory"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: System :: Networking",
]

dependencies = [
    "docker>=6.0.0",
    "requests>=2.28.0",
    "PyYAML>=6.0",
    "scapy>=2.5.0",
    "dpkt>=1.9.8",
    "python-docx>=0.8.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.urls]
Repository = "https://github.com/antonioclim/netENwsl"
Documentation = "https://github.com/antonioclim/netENwsl#readme"
Issues = "https://github.com/antonioclim/netENwsl/issues"

# ═══════════════════════════════════════════════════════════════════════════════
# RUFF CONFIGURATION (Linting and Formatting)
# ═══════════════════════════════════════════════════════════════════════════════

[tool.ruff]
target-version = "py311"
line-length = 100
indent-width = 4

# Directories to check
src = ["src", "scripts", "tests", "formative", "anti_ai"]

# Exclude common non-source directories
exclude = [
    ".git",
    ".github",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "artefacts",
    "pcap/*.pcap",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable these rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]

# Ignore these rules
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Function call in default argument (needed for argparse)
    "B904",   # raise-without-from-inside-except (too strict for beginners)
    "SIM108", # Ternary operator (less readable for students)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
# Tests can use assert and have longer lines
"tests/**/*.py" = ["S101", "E501"]
# Scripts may have different conventions
"scripts/**/*.py" = ["T201"]  # Allow print statements
# Exercise files are educational
"src/exercises/**/*.py" = ["T201", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["src", "scripts", "anti_ai"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

# ═══════════════════════════════════════════════════════════════════════════════
# MYPY CONFIGURATION (Type Checking)
# ═══════════════════════════════════════════════════════════════════════════════

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
show_error_codes = true

# Per-module overrides
[[tool.mypy.overrides]]
module = [
    "scapy.*",
    "dpkt.*",
    "docker.*",
]
ignore_missing_imports = true

# ═══════════════════════════════════════════════════════════════════════════════
# PYTEST CONFIGURATION (Testing)
# ═══════════════════════════════════════════════════════════════════════════════

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Markers for test categorisation
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "docker: marks tests that require Docker",
    "network: marks tests that require network access",
    "integration: marks integration tests",
]

# Command line options
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]

# Filter warnings
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# ═══════════════════════════════════════════════════════════════════════════════
# COVERAGE CONFIGURATION
# ═══════════════════════════════════════════════════════════════════════════════

[tool.coverage.run]
source = ["src", "scripts"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
show_missing = true
fail_under = 70
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"
